Compare expected to measured location
Position error for each move is in the range +/- 0.2 with an occasional 0.5. This is way too high. Time to figure out the controls stuff.

+++ Core/Inc/Dial.h
@@ -15,14 +15,14 @@ namespace std {
 
 class Dial {
 private:
-   uint8_t m_dial_location = 0;
+   float m_dial_location = 0;
    uint8_t m_next_number = 0;
 
 public:
    Dial(); // Constructor
-   int CalculateSteps(enum Direction, uint8_t, uint8_t);
-   uint8_t GetPosition();
-   void UpdatePosition(uint8_t);
+   int CalculateSteps(enum Direction, uint8_t, float);
+   float GetPosition();
+   void UpdatePosition(float);
    ~Dial(); // Destructor
 };
